Update timing logic and allow for more time for bigger JSON files
Updates to the code so Holly Works in missions that have larger JSON files to parse. Other minor edits as well. (#5)

* Update mod.rs

Adjusted the message to reflect what we are doing in the VRM and updated the average contact time from minutes to hours and minutes.

* Update main.rs

Changed the report to only Referrals that are green or yellow dots received over 48 hours ago that have not been successfully contacted or any referral that does not have a contact attempt. Changed the get_average to only look over the past 24 hours.

* Update church.rs

Updated USER_AGENT, increased Client builder timeout from 60 seconds to 240 seconds, and adjusted code with get_person_contact_time to not count for the time between 10:15pm and 6:30am if a referral is received during that time.